Transformations: Moved Guard generation to Rule / Changed PrimeInterpol Strategy to also work via the RewriteEngine
