+++ lib/fcallembed.cpp
@@ -23,7 +23,6 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-#include <cmath>
 #include <flint/ulong_extras.h>
 #include <flint/nmod.h>
 #include <flint/nmod_mat.h>
@@ -37,434 +36,9 @@
 #include "lib/lang.hpp"
 #include "lib/logger.hpp"
 #include "lib/transformations.hpp"
+#include "lib/varstate.hpp"
 
 namespace {
-  /// given a flattend intex recover the access vector needed to create a VarUse Object
-  std::vector<symir::Coef *> unflattenAccess(symir::FunctBuilder *funBuilder, const symir::VarDef *var, size_t flattenedIndex) {
-    if (var->GetType() == symir::SymIR::Type::I32) {
-      Assert(
-        flattenedIndex == 0,
-        "Trying to index int32_to %s which is an (I32, %s) with %ld",
-        var->GetName().c_str(),
-        var->GetBaseType() == symir::SymIR::Type::I32 ? "i32" : (var->GetBaseType() == symir::SymIR::Type::ARRAY ? "ARRAY" : "STRUCT"),
-        flattenedIndex
-      );
-      return {};
-    }
-    std::vector<int32_t> accessVals;
-
-    symir::SymIR::Type type = var->GetType();
-    symir::SymIR::Type baseType = var->GetBaseType();
-    std::string structName =
-        (type == symir::SymIR::Type::STRUCT)
-            ? var->GetStructName()
-            : (baseType == symir::SymIR::Type::STRUCT ? var->GetStructName() : "");
-    std::vector<int32_t> shape = var->IsVector() ? var->GetVecShape() : std::vector<int32_t>{};
-
-    size_t remainingIndex = flattenedIndex;
-
-    // walk down the type tree to generate the access vector
-typeLoop:
-    while (type != symir::SymIR::I32) {
-      switch (type) {
-      case symir::SymIR::ARRAY: {
-        size_t type_size = symir::intSizeOfSymIRType(funBuilder->GetStructs(), baseType, baseType, {}, structName);
-        for (const int32_t dimSize: shape) {
-          type_size *= dimSize;
-        }
-        for (const int32_t dimSize: shape) {
-          type_size /= dimSize;
-          accessVals.push_back(remainingIndex / type_size);
-          remainingIndex %= type_size;
-        }
-        type = baseType;
-        shape = {};
-      } break;
-      case symir::SymIR::STRUCT: {
-        const auto *sDef = funBuilder->FindStruct(structName);
-        Assert(sDef, "Struct %s not found", structName.c_str());
-        size_t fieldIdx = 0;
-        for (const auto &field : sDef->GetFields()) {
-          type = field.type;
-          baseType = field.baseType;
-          if (type == symir::SymIR::Type::STRUCT) {
-            structName = field.structName;
-          } else if (type == symir::SymIR::Type::ARRAY) {
-            shape = field.shape;
-            if (baseType == symir::SymIR::Type::STRUCT) {
-              structName = field.structName;
-            }
-          }
-          size_t field_size = intSizeOfSymIRType(funBuilder->GetStructs(), type, baseType, shape, structName);
-          if (field_size > remainingIndex) {
-            accessVals.push_back(fieldIdx);
-            goto typeLoop;
-          }
-          remainingIndex -= field_size;
-          fieldIdx += 1;
-        }
-      } break;
-      default: Panic("Unknown or Int Type while finding Access Path");
-      }
-    }
-    Assert(
-      remainingIndex == 0,
-      "Var Access has overflown with remainingIndex: %ld, on var: %s, struct: %s, flattendIndex: %ld",
-      remainingIndex, var->GetName().c_str(), var->GetStructName().c_str(), flattenedIndex);
-
-    std::vector<symir::Coef *> access;
-    access.reserve(accessVals.size());
-    // hacky counter to avoid "same named coefficient" error for polynomial coeffs
-    static size_t unique_counter = 0;
-    for (size_t i = 0; i < accessVals.size(); i++) {
-      access.push_back(funBuilder->SymCoef(
-        "poly_var_access_" + var->GetName() + "_" + std::to_string(flattenedIndex) + "_" + std::to_string(i) + "_" + std::to_string(unique_counter),
-        std::to_string(accessVals[i])
-      ));
-    }
-    unique_counter += 1;
-    return access;
-  }
-
-// ==================== PrimeInterpolationHelper ====================
-
-  class PrimeInterpolation {
-  public:
-    PrimeInterpolation(int32_t mod) {
-      Assert(mod <= 46337, "To avoid overflow mod must be less then 46337");
-      Assert(n_is_prime(mod), "PrimeInterpolation requires that mod is prime");
-      nmod_init(&this->mod, mod);
-    }
-  
-    void interpolate(size_t nrVariables, size_t nrIterations, std::vector<int32_t> varState, int32_t target) {
-      Assert(0 <= target && target < static_cast<int32_t>(this->mod.n), "targets (%d) must be in Z_%ld", target, this->mod.n);
-      Assert(nrVariables * nrIterations == varState.size(), "varState is the wrong size");
-      Assert(nrVariables > 0 && nrIterations > 0, "must atleast have one variable and one monomial");
-  
-      Log::Get().OpenSection("Interpolation");
-
-      nmod_mat_t A;
-      nmod_mat_t B;
-      nmod_mat_t X;
-      nmod_mat_init(A, nrIterations + 1, nrIterations + 1, this->mod.n);
-      nmod_mat_init(B, nrIterations + 1, 1, this->mod.n);
-      nmod_mat_init(X, nrIterations + 1, 1, this->mod.n);
-  
-      // Fill B vector
-      for (size_t row = 0; row < nrIterations; row++) {
-        nmod_mat_set_entry(B, row, 0, target);
-      }
-      // Fill the last element of B with a unique new target value to avoid the const polynomial
-      nmod_mat_set_entry(B, nrIterations, 0, nmod_add(static_cast<ulong>(target), 1, this->mod));
-
-
-      int32_t res;
-      for (size_t i = 0; i < 5; i++) { // retry up to 5 times incase the system is unsolvable should rarly happen
-        this->randomizePolynomial(nrVariables, nrIterations);
-  
-        Log::Get().Out() << "Polynomial (" << this->polynomial.size() << "): ";
-        for (size_t i = 0; i < this->polynomial.size(); i++) {
-          Log::Get().Out() << this->polynomial[i] << " ";
-        }
-        Log::Get().Out() << std::endl;
-
-        // set all entries of A
-        for (size_t row = 0; row < nrIterations; row++) {
-          for (size_t col = 0; col < nrIterations; col++) {
-            ulong term = 1;
-            for (uint32_t vari = 0; vari < nrVariables; vari++) {
-              int32_t var = this->reduceMod(varState[row * nrVariables + vari]);
-              ulong deg = static_cast<ulong>(this->polynomial[col * nrVariables + vari]);
-              term = nmod_mul(term, nmod_pow_ui(var, deg, this->mod), this->mod);
-            }
-            nmod_mat_set_entry(A, row, col, term);
-          }
-          nmod_mat_set_entry(A, row, nrIterations, 1);
-        }
-
-        // fill the last row of the A matrix with unique new numbers to avoid the const polynomial
-        std::vector<int32_t> unique_iter =
-          this->findUniqueIteration(nrVariables, nrIterations, varState);
-        for (size_t col = 0; col < nrIterations; col ++) {
-          ulong term = 1;
-          Assert(col * nrVariables < nrIterations * nrVariables, "Array access out of bounds");
-          for (uint32_t i = 0; i < nrVariables; i++) {
-            int32_t var = this->reduceMod(unique_iter[i]);
-            ulong deg = static_cast<ulong>(this->polynomial[col * nrVariables + i]);
-            term = nmod_mul(term, nmod_pow_ui(var, deg, this->mod), this->mod);
-          }
-          nmod_mat_set_entry(A, nrIterations, col, term);
-        }
-        nmod_mat_set_entry(A, nrIterations, nrIterations, 1);
-  
-        Log::Get().Out() << "Interpolation Matrix A["<< nrIterations + 1 << ", " << nrIterations + 1 << "]:" << std::endl;
-        for (size_t row = 0; row < nrIterations + 1; row++) {
-          Log::Get().Out() << "[";
-          for (size_t col = 0; col < nrIterations + 1; col ++) {
-            Log::Get().Out() << nmod_mat_get_entry(A, row, col);
-            if (col != nrIterations) Log::Get().Out() << ", ";
-          }
-          Log::Get().Out() << "]" << std::endl;
-        }
-        Log::Get().Out() << "Rank(A) = " << nmod_mat_rank(A) << std::endl;
-
-        // solve the system
-        res = nmod_mat_can_solve(X, A, B);
-        if (res == 1) break;
-        Log::Get().Out() << "System Unsolvable" << std::endl;
-      }
-      Assert(res == 1, "Unable to solve AX = B");
-  
-      nmod_mat_clear(A);
-      nmod_mat_clear(B);
-  
-      // copy to the output vector
-      this->coeffs.clear();
-      this->coeffs.resize(nrIterations + 1);
-      for (size_t row = 0; row < nrIterations + 1; row++) {
-        this->coeffs[row] = nmod_mat_get_entry(X, row, 0);
-      }
-      nmod_mat_clear(X);
-
-      Log::Get().Out() << std::endl << "Coeffs (" << this->coeffs.size() << "): ";
-      for (size_t i = 0; i < this->coeffs.size(); i++) {
-        Log::Get().Out() << this->coeffs[i] << " ";
-      }
-      Log::Get().Out() << std::endl;
-
-
-      Log::Get().CloseSection();
-    }
-  
-    void interpolate(size_t nrVariables, size_t nrIterations, std::vector<int32_t> varState, std::vector<int32_t> targets) {
-      Assert(targets.size() == nrIterations, "Must have a target value for each iteration");
-      Assert(nrVariables > 0 && nrIterations > 0, "must atleast have one variable and one monomial");
-
-      bool has_unique = false;
-      for (const int32_t target: targets) {
-        Assert(0 <= target && target < static_cast<int32_t>(this->mod.n), "targets must be in Z_%d", static_cast<int32_t>(this->mod.n));
-        has_unique |= targets[0] != target;
-      }
-      Assert(has_unique, "target must have atleast on unique element");
-  
-      this->randomizePolynomial(nrVariables, nrIterations - 1);
-  
-      nmod_mat_t A;
-      nmod_mat_t B;
-      nmod_mat_t X;
-      nmod_mat_init(A, nrIterations, nrIterations, this->mod.n);
-      nmod_mat_init(B, nrIterations, 1, this->mod.n);
-      nmod_mat_init(X, nrIterations, 1, this->mod.n);
-  
-      // set all entries of A
-      for (size_t row = 0; row < nrIterations; row++) {
-        for (size_t col = 0; col < nrIterations - 1; col ++) {
-          ulong term = 1;
-          for (size_t vari = 0; vari < nrVariables; vari++) {
-            int32_t var = this->reduceMod(varState[row * nrVariables + vari]);
-            ulong deg = static_cast<int32_t>(this->polynomial[col * nrVariables + vari]);
-            term = nmod_mul(term, nmod_pow_ui(var, deg, this->mod), this->mod);
-          }
-          nmod_mat_set_entry(A, row, col, term);
-        }
-        nmod_mat_set_entry(A, row, nrIterations - 1, 1);
-      }
-  
-      // build B from m and d
-      for (size_t row = 0; row < nrIterations; row++) {
-        nmod_mat_set_entry(B, row, 0, targets[row]);
-      }
-  
-      // solve the system
-      int32_t res = nmod_mat_can_solve(X, A, B);
-      Assert(res == 1, "Unable to solve AX = B");
-  
-      nmod_mat_clear(A);
-      nmod_mat_clear(B);
-  
-      // build B from m and d
-      this->coeffs.clear();
-      this->coeffs.resize(nrIterations);
-      for (size_t row = 0; row < nrIterations; row++) {
-        this->coeffs[row] = nmod_mat_get_entry(X, row, 0);
-      }
-      nmod_mat_clear(X);
-    }
-  
-    std::vector<int32_t> getPolynomial() {
-      return std::vector(this->polynomial);
-    }
-  
-    std::vector<int32_t> getCoeffs() {
-      return std::vector(this->coeffs);
-    }
-  
-    /// Triggers an assert if the last interpolation does not correctly yield 'target' when evaluated over 'varState'
-    void assertCorrectness(size_t nrVariables, size_t nrIterations, std::vector<int32_t> varState, int32_t target) {
-      Assert(0 <= target && target < static_cast<int32_t>(this->mod.n), "targets (%d) must be in Z_%ld", target, this->mod.n);
-      Assert(nrVariables * nrIterations == varState.size(), "varState is the wrong size");
-      Assert(nrVariables > 0 && nrIterations > 0, "must atleast have one variable and one monomial");
-
-      Assert(this->polynomial.size() == nrVariables * nrIterations, "Interpolation not solved");
-      Assert(this->coeffs.size() == nrIterations + 1, "Interpolation not solved or does not match input");
-
-      for (size_t i = 0; i < nrIterations; i++) {
-        int32_t result = 0;
-        for (size_t m = 0; m < nrIterations; m++) {
-          Assert(0 <= this->coeffs[m] && this->coeffs[m] < static_cast<int32_t>(this->mod.n), "Coef (%d) not in Z_%ld", this->coeffs[m], this->mod.n);
-          int32_t term = this->coeffs[m];
-          for (size_t j = 0; j < nrVariables; j++) {
-            ulong var = this->reduceMod(varState[i * nrVariables + j]);
-            Assert(
-              0 <= this->polynomial[m * nrVariables + j] && this->polynomial[m * nrVariables + j] < static_cast<int32_t>(this->mod.n),
-              "Coef (%d) not in Z_%ld", this->polynomial[m * nrVariables + j], this->mod.n
-            );
-            ulong deg = static_cast<ulong>(this->polynomial[m * nrVariables + j]);
-            term = static_cast<int32_t>(nmod_mul(static_cast<ulong>(term), nmod_pow_ui(var, deg, this->mod), this->mod));
-          }
-          result = static_cast<int32_t>(nmod_add(static_cast<ulong>(result), static_cast<ulong>(term), this->mod));
-        }
-        Assert(
-          0 <= this->coeffs[nrIterations] && this->coeffs[nrIterations] < static_cast<int32_t>(this->mod.n),
-          "Coef (%d) not in Z_%ld", this->coeffs[nrIterations], this->mod.n
-        );
-        result = static_cast<int32_t>(nmod_add(static_cast<ulong>(result), static_cast<ulong>(this->coeffs[nrIterations]), this->mod));
-        Assert(result == target, "interpolation is incorrect!");
-      }
-    }
-
-    /// Triggers an assert if the last interpolation does not correctly yield 'targets' when evaluated over 'varState'
-    void assertCorrectness(size_t nrVariables, size_t nrIterations, std::vector<int32_t> varState, std::vector<int32_t> targets) {
-      bool has_unique = false;
-      for (const int32_t target: targets) {
-        Assert(0 <= target && target < static_cast<int32_t>(this->mod.n), "targets must be in Z_%d", static_cast<int32_t>(this->mod.n));
-        has_unique |= targets[0] != target;
-      }
-      Assert(has_unique, "target must have atleast on unique element");
-      Assert(nrVariables * nrIterations == varState.size(), "varState is the wrong size");
-      Assert(nrVariables > 0 && nrIterations > 0, "must atleast have one variable and one monomial");
-
-      Assert(this->polynomial.size() == nrVariables * (nrIterations - 1), "Interpolation not solved");
-      Assert(this->coeffs.size() == nrIterations, "Interpolation not solved or does not match input");
-
-      for (size_t i = 0; i < nrIterations; i++) {
-        int32_t result = 0;
-        for (size_t m = 0; m < nrIterations - 1; m++) {
-          Assert(0 <= this->coeffs[m] && this->coeffs[m] < static_cast<int32_t>(this->mod.n), "Coef (%d) not in Z_%ld", this->coeffs[m], this->mod.n);
-          int32_t term = this->coeffs[m];
-          for (size_t j = 0; j < nrVariables; j++) {
-            ulong var = this->reduceMod(varState[i * nrVariables + j]);
-            Assert(
-              0 <= this->polynomial[m * nrVariables + j] && this->polynomial[m * nrVariables + j] < static_cast<int32_t>(this->mod.n),
-              "Coef (%d) not in Z_%ld", this->polynomial[m * nrVariables + j], this->mod.n
-            );
-            ulong deg = static_cast<ulong>(this->polynomial[m * nrVariables + j]);
-            term = static_cast<int32_t>(nmod_mul(static_cast<ulong>(term), nmod_pow_ui(var, deg, this->mod), this->mod));
-          }
-          result = static_cast<int32_t>(nmod_add(static_cast<ulong>(result), static_cast<ulong>(term), this->mod));
-        }
-        Assert(
-          0 <= this->coeffs[nrIterations - 1] && this->coeffs[nrIterations - 1] < static_cast<int32_t>(this->mod.n),
-          "Coef (%d) not in Z_%ld", this->coeffs[nrIterations - 1], this->mod.n
-        );
-        result = static_cast<int32_t>(nmod_add(static_cast<ulong>(result), static_cast<ulong>(this->coeffs[nrIterations - 1]), this->mod));
-        Assert(result == targets[i], "interpolation is incorrect!");
-      }
-    }
-
-  private:
-  
-    /// Find a new unused iteration according to varState
-    std::vector<int32_t> findUniqueIteration(size_t nrVariables, size_t nrIterations, std::vector<int32_t> varState) {
-      std::vector<int32_t> unique;
-      std::vector<int32_t> currVarState;
-      unique.reserve(nrVariables);
-      currVarState.resize(nrIterations);
-      for (size_t i = 0; i < nrVariables; i++) {
-        // Copy and sort to make it easy to find new unique value
-        for (size_t j = 0; j < nrIterations; j++) {
-          int32_t var = varState[j * nrVariables + i];
-          currVarState[j] = this->reduceMod(var);
-        }
-        std::sort(currVarState.begin(), currVarState.end());
-  
-        // To actually find a unique element we look for the largest difference between to successive elements.
-        // This is to minimize the chance the AX = B is unsolvable
-        int32_t last = 0;
-        int32_t largest_range = 0;
-        int32_t candidate = -1;
-        for (size_t k = 0; k < nrIterations; k++) {
-          // if the next value is not the successor of - or equal to the last one there must be a unique value inbetween
-          if (currVarState[k] != static_cast<int32_t>((last + 1) % mod.n)
-              && currVarState[k] - 1 - (last + 1) >= largest_range) {
-            largest_range = currVarState[k] - 1 - (last + 1);
-            candidate = Random::Get().Uniform(last + 1, currVarState[k] - 1)();
-          }
-          last = currVarState[k];
-        }
-        if (static_cast<int32_t>(mod.n - 1) - currVarState.back() >= largest_range) {
-          largest_range = static_cast<int32_t>(mod.n - 1) - currVarState.back();
-          candidate = Random::Get().Uniform(currVarState.back() + 1, static_cast<int32_t>(mod.n - 1))();
-        }
-        if (largest_range > 0) unique.push_back(candidate);
-        else Panic("unable for find a unique element");
-      }
-      Assert(unique.size() == nrVariables, "unique array not of the right size");
-      return unique;
-    }
-    
-    /// see https://en.wikipedia.org/wiki/Fisher%E2%80%93Yates_shuffle
-    void shuffel(const size_t n, int32_t *arr) {
-      for (size_t i = n - 1; i >= 1; i--) {
-        size_t r = rand() % (i + 1);
-        int32_t temp = arr[r];
-        arr[r] = arr[i];
-        arr[i] = temp;
-      }
-    }
-    
-    // This is a quite biased in its selection. TODO: Are the more uniform / fair algorithms for finding a random polynomial fast?
-    // Monomial ordering: https://people.math.sc.edu/Burkardt/c_src/monomial/monomial.html
-    /// Samples a new random polynomial
-    void randomizePolynomial(
-      const size_t nrVariables,
-      const size_t nrMonomials
-    ) {
-      Assert(nrVariables > 0 && nrMonomials > 0, "must atleast have one variable and one monomial");
-      this->polynomial.clear();
-      this->polynomial.resize(nrVariables * nrMonomials);
-      for (size_t d = nrMonomials; d > 0; d--) {
-        Assert((nrMonomials - d) * nrVariables < nrMonomials * nrVariables, "Array out of bounds");
-        int32_t *monomial = &this->polynomial[(nrMonomials - d) * nrVariables];
-        int32_t upper = d;
-        for (size_t j = 0; j < nrVariables - 1; j++) {
-          monomial[j] = abs(Random::Get().Binomial(upper * 2, 0.5)() - upper);
-          upper -= monomial[j];
-          if (upper == 0) break;
-        }
-        monomial[nrVariables - 1] = upper;
-        this->shuffel(nrVariables, monomial);
-      }
-    }
-
-    /// Returns the (mathematical) modulo of 'var' (e.g. 'var' mod 'this->mod.n')
-    ulong reduceMod(int32_t var) {
-      ulong res;
-      if (var < 0) {
-        NMOD_RED(res, static_cast<ulong>(-static_cast<int64_t>(var)), this->mod);
-        res = res != 0 ? this->mod.n - res : res;
-      } else {
-        NMOD_RED(res, static_cast<ulong>(var), this->mod);
-      }
-      Assert(res < this->mod.n, "reduceMod has produced 'res' not in Z_%ld", this->mod.n);
-      return res;
-    }
-  private:
-    nmod_t mod;
-    std::vector<int32_t> polynomial;
-    std::vector<int32_t> coeffs;
-  };
-
   void retargetBlock(symir::FunctBuilder *funBd, const symir::Block *blk, std::map<const std::string, symir::BlockBuilder *> argBlocks) {
     const symir::Target *target = blk->GetTarget();
     if (target == nullptr) return;
@@ -511,7 +85,6 @@ typeLoop:
       Panic("Each block must have a target but target IR has Id: %d", target->GetIRId());
     }
   }
-
 } // namespace
 
 // ==================== FCallStrategy Base Implementations ====================
@@ -579,95 +152,6 @@ const symir::VarDef *FCallStrategy::getUnusedAssignVar(symir::FunctBuilder *funB
   return loc;
 }
 
-void FCallStrategy::appendVarState(VariableStateQuery *varStateQuery, size_t blockIndex, size_t stmtIndex) {
-  // get the current variable state for all variables in scope
-  auto varStatePair = varStateQuery->query(blockIndex, stmtIndex);
-  size_t currVarStateSize;
-  if (this->nrVariables == 0) {
-    this->nrVariables = varStatePair.first;
-    this->varState = varStatePair.second;
-    currVarStateSize = this->varState.size();
-  } else {
-    Assert(this->nrVariables == varStatePair.first, "Different Queries have different nrVariables");
-    this->varState.reserve(this->varState.size() + varStatePair.second.size());
-    currVarStateSize = varStatePair.second.size();
-    for (size_t i = 0; i < currVarStateSize; i++) {
-      this->varState.push_back(varStatePair.second[i]);
-    }
-  }
-  Assert(this->varState.size() % this->nrVariables == 0, "nrVariables * nr_iteration == varState.size must hold");
-  this->nrIterations += currVarStateSize / this->nrVariables;
-  
-  Log::Get().Out() << "Variables State: (nrVariables: " << this->nrVariables << "), (nrIterations: " << this->nrIterations << ")" << std::endl;
-  for (size_t i = 0; i < this->nrIterations; i++) {
-    Log::Get().Out() << "Iteration " << i << ": ";
-    for (size_t j = 0; j < this->nrVariables; j++) {
-      Log::Get().Out() << this->varState[i * nrVariables + j];
-      if (j == this->nrVariables - 1) {
-        Log::Get().Out() << std::endl;
-      } else {
-        Log::Get().Out() << ", ";
-      }
-    }
-  }
-  Log::Get().Out() << std::endl;
-}
-
-void FCallStrategy::randomlyFilterVarState(symir::FunctBuilder *funBd) {
-  auto randDouble = Random::Get().UniformReal();
-  // randomly select a subset of variables and find there VarDef
-  this->filteredVars.clear();
-  this->filteredAccesses.clear();
-  std::vector<size_t> indices;
-  size_t lastVarStartIndex = 0;
-  for (size_t i = 0; i < nrVariables; i++) {
-    if (this->varMap.contains(i)) lastVarStartIndex = i;
-    if (indices.size() > 0 && randDouble() > 1 - GlobalOptions::Get().VariableTakeProba) continue;
-    indices.push_back(i);
-    const symir::VarDef *var = funBd->FindVar(this->varMap[lastVarStartIndex]);
-    this->filteredVars.push_back(var);
-    this->filteredAccesses.push_back(unflattenAccess(funBd, var, i - lastVarStartIndex));
-  }
-  this->filteredNrVariables = this->filteredVars.size();
-
-  Log::Get().Out() << "Using Variables (" << this->filteredNrVariables << "): ";
-  for (size_t i = 0; i < this->filteredNrVariables; i++) {
-    Log::Get().Out() << this->filteredVars[i]->GetName();
-    if (this->filteredVars[i]->GetType() != symir::SymIR::I32) {
-      Log::Get().Out() << "[";
-      for (size_t j = 0; j < this->filteredAccesses[i].size() - 1; j++) {
-        Log::Get().Out() << this->filteredAccesses[i][j]->GetI32Value() << ", ";
-      }
-      Log::Get().Out() << this->filteredAccesses[i].back()->GetI32Value() << "]";
-    }
-    Log::Get().Out() << "(" << indices[i] << "): [";
-    for (size_t k = 0; k < this->nrIterations; k++) {
-      Log::Get().Out() << this->varState[k * nrVariables + indices[i]];
-      if (k == this->nrIterations - 1) {
-        Log::Get().Out() << "]";
-      } else {
-        Log::Get().Out() << ", ";
-      }
-    }
-    if (i == this->filteredVars.size() - 1) {
-      Log::Get().Out() << std::endl;
-    } else {
-      Log::Get().Out() << ", ";
-    }
-  }
-  
-  // Filter varState
-  this->filteredVarState.clear();
-  this->filteredVarState.reserve(this->nrIterations * this->filteredNrVariables);
-  for (size_t i = 0; i < this->nrIterations; i++) {
-    for (size_t j = 0; j < this->filteredNrVariables; j++) {
-      this->filteredVarState.push_back(this->varState[i * this->nrVariables + indices[j]]);
-    }
-  }
-}
-
-void smartlyFilterVarState(const symir::FunctBuilder *funBd) { Panic("TODO: Not yet implemented"); }
-
 // ==================== FCallEmbedder Base Implementations ====================
 FCallEmbedder::FCallEmbedder(symir::Funct *const host): host(host) {
   Assert(
@@ -722,73 +206,6 @@ void FCallEmbedder::markMutated(symir::Coef *c) {
   it->second = true;
 }
 
-// ==================== Guard Generator Implementations ====================
-
-symir::BlockBuilder::TermID ModInterpGuardStrategy::addGuard(
-  symir::FunctBuilder *funBd,
-  symir::BlockBuilder *blockBd,
-  size_t nrVariables,
-  size_t nrIterations,
-  std::vector<const symir::VarDef *> variables,
-  std::vector<std::vector<symir::Coef *>> accesses,
-  std::vector<int32_t> varState,
-  const symir::Term * targetTerm,
-  size_t nthGuard
-) const {
-  Assert(targetTerm->GetOp() == symir::Term::OP_CST, "target must be a constant");
-  Assert(targetTerm->GetCoef()->IsSolved(), "target coef must be solved");
-
-  int32_t prime = 46337; //TODO: maybe allow users to choose 2 <= P <= 46337 via a global?
-
-  PrimeInterpolation interpolGen = PrimeInterpolation(prime); 
-
-  std::vector<const symir::Param *> params = funBd->GetParams();
-  std::vector<const symir::Local *> locals = funBd->GetLocals();
-
-  auto randDouble = Random::Get().UniformReal();
-  auto randTarget = Random::Get().Uniform(0, prime - 1);
-
-  // We need a target value in Z_p to achive this we choose one randomly and then figure out how to correct for it.
-  int32_t target = targetTerm->GetCoef()->GetI32Value();
-  int32_t interpolTarget;
-  if (target == INT32_MIN) {
-    // avoid the div by 0 case of the else stmt
-    interpolTarget = 0;
-  } else {
-    interpolTarget = randTarget() % static_cast<int32_t>((-static_cast<int64_t>(INT32_MIN)) + target);
-  }
-
-  Log::Get().Out() << "Target: " << target << ", Interpolation Target: " << interpolTarget << std::endl;
-
-  interpolGen.interpolate(nrVariables, nrIterations, varState, interpolTarget);
-  std::vector<int32_t> polynomial = interpolGen.getPolynomial();
-  std::vector<int32_t> coeffsVals = interpolGen.getCoeffs();
-  interpolGen.assertCorrectness(nrVariables, nrIterations, varState, interpolTarget);
-
-  // we need coeff object to hand to the builder
-  std::vector<symir::Coef *> coeffs{};
-  coeffs.reserve(coeffsVals.size());
-  for (const int32_t c : coeffsVals) {
-    coeffs.push_back(funBd->SymI32Const(c));
-  }
-
-  const symir::VarDef *loc = funBd->FindVar("guard_" + std::to_string(nthGuard));
-  if (loc == nullptr) {
-    loc = funBd->SymScaLocal("guard_" + std::to_string(nthGuard), nullptr);
-  } 
-
-  blockBd->CommitStmtAt(
-    blockBd->SymModAssStmt(
-      loc,
-      blockBd->SymModExpr(coeffs, variables, accesses, polynomial, prime),
-      {}
-    ),
-    0
-  );
-  Log::Get().Out() << std::endl;
-  return blockBd->SymAddTerm(funBd->SymI32Const(target - interpolTarget), loc);
-}
-
 // ==================== LiteralFCallStrategy Implementations ====================
 std::string LiteralFCallStrategy::generateCall() {
   Assert(this->guest, "guest is not initialized");
@@ -841,38 +258,23 @@ void PrimeInterpFCallStrategy::generatePreamble(
   Assert(this->fina, "fina is not initialized");
   Assert(blockIndex < funBd->GetBlocks().size(), "index %ld is Out of bound for blocks", blockIndex);
 
-  Log::Get().OpenSection("generatePreable");
-
-  // clear temporaries
-  this->argVars.clear();
-  this->varState.clear();
-  this->varMap.clear();
-  this->filteredVarState.clear();
-  this->filteredVars.clear();
-  this->filteredAccesses.clear();
-  this->nrVariables = 0;
-  this->nrIterations = 0;
-  this->filteredNrVariables = 0;
+  Log::Get().OpenSection("RevOptFCallStrategy::generatePreable");
 
   if (this->nrBlocks == 0) this->setMaxNrBlocks(funBd->GetBlocks().size());
 
-  int32_t prime = 46337; //TODO: maybe allow users to choose 2 <= P <= 46337 via a global?
-  PrimeInterpolation interpolGen = PrimeInterpolation(prime); 
-
-  std::vector<const symir::Param *> params = funBd->GetParams();
-  std::vector<const symir::Local *> locals = funBd->GetLocals();
-
   const symir::Block *targetBlock= funBd->GetBlocks()[blockIndex];
-  Log::Get().Out() << "Targeting Block: " << targetBlock->GetLabel() << ", " << stmtIndex << "-th Statement" << std::endl;
-  symir::BlockBuilder *blockBd = symir::BlockCopier(funBd, targetBlock).CopyAsBuilder();
-
-  this->varMap = varStateQueries[0]->GetVarMap();
-  for (size_t i = 0; i < varStateQueries.size(); i++) {
-    this->appendVarState(varStateQueries[i], blockIndex, stmtIndex);
+  const std::string targetLabel = targetBlock->GetLabel();
+  Log::Get().Out() << "targeting block: " << targetLabel << "at stmt: " << stmtIndex << std::endl;
+  symir::BlockBuilder *headerBlockBd;
+  if (!argBlocks.contains(targetLabel)) {
+    Log::Get().Out() << "Creating new block " << targetBlock->GetLabel() + "_header" << std::endl;
+    this->argBlocks[targetLabel] = funBd->OpenBlock(targetBlock->GetLabel() + "_header");
+  } else {
+    Log::Get().Out() << "Reusing block " << targetBlock->GetLabel() + "_header" << std::endl;
   }
+  headerBlockBd = this->argBlocks[targetBlock->GetLabel()];
 
   auto randDouble = Random::Get().UniformReal();
-  auto randTarget = Random::Get().Uniform(0, prime - 1);
   size_t flattIndex = 0;
   for (size_t initIdx = 0; initIdx < this->init->size(); initIdx++) {
     const auto &arg = (*this->init)[initIdx];
@@ -880,55 +282,27 @@ void PrimeInterpFCallStrategy::generatePreamble(
       flattIndex += 1;
       if (randDouble() <= 1 - GlobalOptions::Get().InitReplaceProba) continue;
 
-      Log::Get().Out() << "Replacing the " << flattIndex << "-th argument" << std::endl;
+      Log::Get().Out() << "Replacing the flattend " << flattIndex << "-th argument" << std::endl;
 
-      this->randomlyFilterVarState(funBd);
+      const symir::VarDef *loc = this->getUnusedAssignVar(funBd, blockIndex, 0);
+      int val = arg.IsScalar() ? arg.GetValue() : arg.GetValue(argIdx);
 
-      // We need a target value in Z_p to achive this we choose one randomly and then figure out how to correct for it.
-      int32_t target = arg.IsScalar() ? arg.GetValue() : arg.GetValue(argIdx);
-      int32_t interpolTarget;
-      if (target == INT32_MIN) {
-        // avoid the div by 0 case of the else stmt
-        interpolTarget = 0;
-      } else {
-        interpolTarget = randTarget() % static_cast<int32_t>((-static_cast<int64_t>(INT32_MIN)) + target);
-      }
-
-      Log::Get().Out() << "Target: " << target << ", Interpolation Target: " << interpolTarget << std::endl;
-
-      interpolGen.interpolate(this->filteredNrVariables, this->nrIterations, this->filteredVarState, interpolTarget);
-      std::vector<int32_t> polynomial = interpolGen.getPolynomial();
-      std::vector<int32_t> coeffsVals = interpolGen.getCoeffs();
-      interpolGen.assertCorrectness(filteredNrVariables, nrIterations, filteredVarState, interpolTarget);
-
-      // we need coeff object to hand to the builder
-      std::vector<symir::Coef *> coeffs{};
-      coeffs.reserve(coeffsVals.size());
-      for (const int32_t c : coeffsVals) {
-        coeffs.push_back(funBd->SymI32Const(c));
-      }
-
-      const symir::VarDef *loc = this->getUnusedAssignVar(funBd, blockIndex, stmtIndex);
-
-      blockBd->CommitStmtAtAssign(
-        blockBd->SymModAssStmt(
-          loc,
-          blockBd->SymModExpr(
-            coeffs,
-            this->filteredVars,
-            this->filteredAccesses,
-            polynomial,
-            prime
-          ),
-          {}
-        ), 
-        stmtIndex
+      Log::Get().Out() << loc->GetName() << " <- " << val << std::endl;
+      headerBlockBd->CommitStmt(
+        headerBlockBd->SymAssStmt(
+          loc, 
+          headerBlockBd->SymAddExpr({
+            headerBlockBd->SymCstTerm(
+              funBd->SymI32Const(val),
+              nullptr
+            )
+          })
+        )
       );
-      this->argVars[flattIndex - 1] = std::make_pair(loc->GetName(), target - interpolTarget);
+      this->argVars[flattIndex - 1] = std::make_pair(loc->GetName(), 0);
       Log::Get().Out() << std::endl;
     }
   }
-  funBd->ReplaceOrCloseBlock(blockBd);
   Log::Get().CloseSection();
 }
 
@@ -976,6 +350,71 @@ std::string PrimeInterpFCallStrategy::generateCall() {
   } else {
     return "(int) ((long long)" + chk_call + " + " + std::to_string(diff) + "L)";
   }
+}
+
+void PrimeInterpFCallStrategy::finalize(std::vector<VariableStateQuery *> varStateQueries, symir::FunctBuilder *funBd) {
+  // needs variable state
+  Log::Get().OpenSection("PrimeInterpFCallStrategy::finalize for " + funBd->GetName());
+
+  // delete all header blocks that do not contain any stmt
+  for (auto it = this->argBlocks.cbegin(); it != this->argBlocks.cend();) {
+    if (it->second->GetNumberOfCommitedStmt() == 0) {
+      it = this->argBlocks.erase(it);
+    } else {
+      ++it;
+    }
+  }
+
+  // loop through all blocks and if the target has a header change their target to the header.
+  for (auto &blk : funBd->GetBlocks()) {
+    retargetBlock(funBd, blk, this->argBlocks);
+  }
+
+  std::map<std::string, size_t> labelToIdx;
+  size_t idx = 0;
+  for (auto &blk : funBd->GetBlocks()) {
+    labelToIdx[blk->GetLabel()] = idx++;
+  }
+
+  Assert(varStateQueries.size() > 0, "must have atleast on Variable State Query");
+
+  for (auto const &[blkLabel, headerBlockBd] : this->argBlocks) {
+
+    // set the target of the header to the block
+    const symir::Block* blk = funBd->FindBlock(blkLabel);
+    Assert(blk != nullptr, "Unable to find block %s in %s", blkLabel.c_str(), funBd->GetName().c_str());
+    Assert(blk->GetLabel() == blkLabel, "function builders internal Map must be broken (%s != %s)", blk->GetLabel().c_str(), blkLabel.c_str());
+    headerBlockBd->SymGoto(blkLabel);
+    Log::Get().Out() << "Setting target of header block " << headerBlockBd->GetLabel()
+                     << " to " << blkLabel << std::endl;
+    
+    // Get the Variable State or the original function
+    struct VariableState totalVariableState;
+    totalVariableState.nrVariables = 0;
+    for (auto &varStateQuery : varStateQueries) {
+      struct VariableState vs = varStateQuery->query(labelToIdx[blkLabel], 0);
+      if (totalVariableState.nrVariables == 0) {
+        totalVariableState.nrVariables = vs.nrVariables;
+        totalVariableState.varMap = vs.varMap;
+      }
+      Assert(totalVariableState.nrVariables == vs.nrVariables, "Query returns more or less variables then there are in varMap");
+      for (size_t i = 0; i < vs.varState.size(); i++) {
+        totalVariableState.varState.push_back(vs.varState[i]);
+      }
+    }
+
+    std::vector<symir::BlockBuilder *> headerBlockBds = { headerBlockBd };
+    transformations::primitive::Guard rule = transformations::primitive::Guard();
+    this->rewriteEngine.runAsPass(funBd, headerBlockBds, totalVariableState, rule);
+
+    for (auto blockBd : headerBlockBds) {
+      funBd->CloseBlockAt(blockBd, blk);
+    }
+  }
+
+  // avoid causing problems by calling this function twice;
+  this->argBlocks.clear();
+  Log::Get().CloseSection();
 }
 
 void RevOptFCallStrategy::generatePreamble(
@@ -1106,7 +545,6 @@ void RevOptFCallStrategy::finalize(std::vector<VariableStateQuery *> varStateQue
 
   Assert(varStateQueries.size() > 0, "must have atleast on Variable State Query");
 
-  this->varMap = varStateQueries[0]->GetVarMap();
   for (auto const &[blkLabel, headerBlockBd] : this->argBlocks) {
 
     // set the target of the header to the block
@@ -1117,8 +555,23 @@ void RevOptFCallStrategy::finalize(std::vector<VariableStateQuery *> varStateQue
     Log::Get().Out() << "Setting target of header block " << headerBlockBd->GetLabel()
                      << " to " << blkLabel << std::endl;
 
+    // Get the Variable State or the original function
+    struct VariableState totalVariableState;
+    totalVariableState.nrVariables = 0;
+    for (auto &varStateQuery : varStateQueries) {
+      struct VariableState vs = varStateQuery->query(labelToIdx[blkLabel], 0);
+      if (totalVariableState.nrVariables == 0) {
+        totalVariableState.nrVariables = vs.nrVariables;
+        totalVariableState.varMap = vs.varMap;
+      }
+      Assert(totalVariableState.nrVariables == vs.nrVariables, "Query returns more or less variables then there are in varMap");
+      for (size_t i = 0; i < vs.varState.size(); i++) {
+        totalVariableState.varState.push_back(vs.varState[i]);
+      }
+    }
+
     std::vector<symir::BlockBuilder *> headerBlockBds = { headerBlockBd };
-    this->rewriteEngine.run(funBd, headerBlockBds, 100);
+    this->rewriteEngine.run(funBd, headerBlockBds, totalVariableState, 100);
 
     for (auto blockBd : headerBlockBds) {
       funBd->CloseBlockAt(blockBd, blk);
